# Conflicts:
#	HADES/App_GlobalResources/Strings.en-US.resx
#	HADES/App_GlobalResources/Strings.resx

+++ HADES/Models/AppConfig.cs
@@ -48,7 +48,7 @@ namespace HADES.Models
         [Column("ACF_log_max_file_size")]
         public int LogMaxFileSize { get; set; }
 
-        public int LogTotalMaxSize => LogMaxFileSize * LogDeleteFrequency;
+        public uint LogTotalMaxSize => (uint)(LogMaxFileSize * LogDeleteFrequency);
 
         [ForeignKey("ActiveDirectory")]
         [Column("ACF_ADR_id")]
